+++ pkg/devfile/consts/consts.go
@@ -1,0 +1,5 @@
+package consts
+
+// UriFolder is the sub-directory where non inlined kubernetes component
+// definitions are saved
+const UriFolder = "kubernetes"
